refactor: report api 분리 (#81)
