+++ src/entities/report/api/report-service.ts
@@ -1,0 +1,19 @@
+import { Client } from '@/shared/api/client'
+import { ENDPOINT } from '@/shared/model/constants'
+import { BugReportPayload, ReportService } from '../model/report-interface'
+
+class ReportApiService implements ReportService {
+  readonly #client: Client
+
+  constructor() {
+    this.#client = Client.Instance
+  }
+
+  async sendReport(payload: BugReportPayload) {
+    return this.#client.post(ENDPOINT.BUG_REPORT, payload)
+  }
+}
+
+const ReportApi = new ReportApiService()
+
+export { ReportApi }
