+++ src/entities/report/api/send-report-action.ts
@@ -1,11 +1,12 @@
 'use server'
 
 import axios from 'axios'
-import { SpellerApi, type BugReportPayload } from '@/entities/speller'
+import { type BugReportPayload } from '@/entities/speller'
+import { ReportApi } from './report-service'
 
 export const sendReportAction = async (payload: BugReportPayload) => {
   try {
-    await SpellerApi.sendReport(payload)
+    await ReportApi.sendReport(payload)
   } catch (error: unknown) {
     if (axios.isAxiosError(error)) {
       throw new Error(error.message)
